Struct rustboost::boosting::Boosting

pub struct Boosting { /* fields omitted */ }

Entry class of the boosting algorithm. It contains two functions, one for starting the network communication, the other for starting the training procedure.

Methods

impl Boosting

pub fn new(
training_loader: BufferLoader,
range: Range<usize>,
max_sample_size: usize,
max_bin_size: usize,
default_gamma: f32
) -> Boosting

Create a boosting training class.

training_loader: the double-buffered data loader that provides examples to the algorithm.

range: the range of the feature dimensions that the weak rules would be selected from. In most cases, if the RustBoost is running on a single worker, range is equal to the 0..feature_size; if it is running over multiple workers, it might be a subset of the full feature set.

max_sample_size: the number of examples to scan for determining the percentiles for the features.

max_bin_size: the size of the percentiles to generate on each feature dimension.

default_gamma: the initial value of the edge gamma of the candidate valid weak rules.

pub fn enable_network(
&mut self,
name: String,
remote_ips: &Vec<String>,
port: u16
)

Enable network communication. name is the name of this worker, which can be arbitrary and is only used for debugging purpose. remote_ips is the vector of IPs of neighbor workers. port is the port number that used for network communication.

pub fn training(&mut self, num_iterations: usize, max_trials_before_shrink: u32)

Start training the boosting algorithm.

num_iterations: the number of boosting iteration. If it equals to 0, then the algorithm runs indefinitely.

max_trials_before_shrink: if cannot find any valid weak rules after scanning max_trials_before_shrink number of examples, shrinking the value of the targetting edge gamma of the weak rule.