Struct rustboost::learner::Learner

pub struct Learner { /* fields omitted */ }

Statisitics of all weak rules that are being evaluated. The objective of Learner is to find a weak rule that satisfies the condition of the stopping rule.

Methods

impl Learner

pub fn new(default_gamma: f32, bins: Vec<Bins>, range: &Range<usize>) -> Learner

Create a Learner that search for valid weak rules. default_gamma is the initial value of the edge gamma. bins is vectors of the all thresholds on all candidate features for generating weak rules.

range is the range of candidate features for generating weak rules. In most cases, if the algorithm is running on a single worker, range is 0..num_of_features; if the algorithm is running on multiple workers, range is a subset of the feature set.

pub fn reset(&mut self)

Reset the statistics of all candidate weak rules, but leave the targetting gamma unchanged.

pub fn get_count(&self) -> usize

Return the number of examples scanned so far since detecting the last valid weak rule or since last reset.

pub fn get_gamma(&self) -> f32

Return current targetting edge gamma

pub fn reset_all(&mut self)

Reset the targetting gamma to initial_gamma, and reset all statistics of all candidate weak rules.

pub fn shrink_target(&mut self)

Shrinking the value of the targetting edge gamma of the valid weak rule

pub fn update(&mut self, data: &[ExampleInSampleSet], weights: &Vec<f32>)

Update the statistics of all candidate weak rules using current batch of training examples.

pub fn get_new_weak_rule(&mut self) -> &Option<WeakRule>

Return the valid weak rule if one is found, otherwise return None.